Move extract tasks to ingestion.
