+++ pipeline/matching_pipeline.py
@@ -12,82 +12,22 @@ from firebase_admin import credentials, db
 import prefect
 from prefect import Flow, Parameter, task
 from prefect.tasks.secrets import PrefectSecret
+from pipeline.ingestion.users import extract_users
+from pipeline.ingestion.matches import extract_recent_matches
 from pipeline.matching import (
-    Match,
-    User,
     best_effort_minimize_repeat_matches,
     get_matches,
     validate_matches,
 )
-from pipeline.utils.matches import (
+from pipeline.schema.user import User
+from pipeline.schema.match import Match
+from pipeline.utils.constants import MS
+from pipeline.utils.firebase import initialize_firebase_for_prefect
+from pipeline.utils.release import (
     generate_keys,
     from_release_tag,
     to_release_tag,
 )
-
-
-MS = 1000
-
-
-@task
-def initialize_firebase_for_prefect(database_url: str, admin_credentials: str):
-    # Parse and write credentials to file
-    credentials_outfile = "prefect-credentials.json"
-    raw = str(admin_credentials).replace("'", '"').replace("^", " ")
-    parsed = json.loads(raw)
-    content = json.dumps(parsed, indent=4)
-    with open(credentials_outfile, "w+") as file:
-        file.write(content)
-
-    # Initialize Firebase Admin service account
-    cred = credentials.Certificate(credentials_outfile)
-    firebase_admin.initialize_app(cred, {"databaseURL": database_url})
-
-    # After initialization, delete raw credentials file
-    if os.path.exists(credentials_outfile):
-        os.remove(credentials_outfile)
-    return db
-
-
-@task
-def extract_users(db, community: str) -> pd.DataFrame:
-    logger = prefect.context.get("logger")
-    all_users = db.reference("users").get()
-    df_users = pd.DataFrame(all_users.values())
-    # Filter out users who have not logged in
-    df_users = df_users[df_users.latestLogin.notna()]
-    logger.info("Returned {} rows, {} cols.".format(*df_users.shape))
-    return df_users
-
-
-@task
-def extract_recent_matches(db, community: str) -> pd.DataFrame:
-    logger = prefect.context.get("logger")
-    matches_ref = db.reference(f"matches/{community}")
-    user_match_records = matches_ref.get()
-    logger.info(f"Extract {len(user_match_records)} user-match records.")
-
-    # Convert user-match records to match records
-    seen_match_ids = set()
-    recent_match_records = []
-    for record in user_match_records.values():
-        match_key = record["id"]
-        if match_key in seen_match_ids:
-            continue
-        seen_match_ids.add(match_key)
-        match = Match(
-            users={uid for uid in record["participants"]},
-            community=community,
-            release=from_release_tag(record["release_tag"]),
-            key=match_key,
-            title=record["title"],
-        )
-        recent_match_records.append(match)
-    logger.info(f"Converted to {len(recent_match_records)} match records.")
-
-    df_recent_matches = pd.DataFrame(recent_match_records)
-    logger.info("Returned {} rows, {} cols.".format(*df_recent_matches.shape))
-    return df_recent_matches
 
 
 @task
