+++ pipeline/schema/match.py
@@ -1,0 +1,11 @@
+import datetime
+from pipeline.schema.user import UserId
+
+
+@dataclass
+class Match:
+    users: Set[UserId]
+    community: Optional[str] = None
+    release: Optional[datetime.datetime] = None
+    key: Optional[str] = None
+    title: Optional[str] = None
