+++ pipeline/schema/user.py
@@ -1,0 +1,7 @@
+UserId = str
+
+
+@dataclass
+class User:
+    uid: UserId
+    displayName: str
