+++ pipeline/utils/constants.py
@@ -1,0 +1,1 @@
+MS = 1000
